requirements.txt
discord.py == 2.6.4
python-dotenv>=1
beautifulsoup4
pdfkit
aiohttp
google-api-python-client
dateparser
# plus besoin de requests-htmlc (Playwright viré)
git+https://github.com/Sergeileduc/python-web-tools.git@main
# -e "C:\Users\serge\Dev\Packages\Python\python-web-tools"
git+https://github.com/Sergeileduc/lemonde-sl.git
playwright
